Finished admin system and fixed authorization

+++ app/api/class/[id]/route.ts
@@ -2,8 +2,32 @@ import { auth } from "@/lib/auth";
 import { prisma } from "@/lib/prisma";
 import { NextResponse } from "next/server";
 
-export const DELETE = auth(async (_, { params }) => {
-  const { id } = params as { id: string };;
+export const DELETE = auth(async (req, { params }) => {
+  const { id } = params as { id: string };
+
+  if (!req.auth?.user)
+    return NextResponse.json(
+      {
+        message: "Unauthorized",
+      },
+      {
+        status: 401,
+      }
+    );
+
+  const user = await prisma.user.findUnique({
+    where: { id: req.auth.user.id },
+  });
+
+  if (user?.role !== "ADMIN")
+    return NextResponse.json(
+      {
+        message: "Unauthorized",
+      },
+      {
+        status: 401,
+      }
+    );
 
   await prisma.class.delete({
     where: { id },
